refactor: consolidate authentication logic by merging LoginController into AuthController
